make daily summary time window more sophisticated, rename DailySummaryAll to ProcessAllDailySummary, move view render logic to its own package
