+++ server/command/view_calendar.go
@@ -5,13 +5,20 @@ package command
 
 import (
 	"time"
+
+	"github.com/mattermost/mattermost-plugin-mscalendar/server/mscalendar/views"
 )
 
 func (c *Command) viewCalendar(parameters ...string) (string, error) {
+	tz, err := c.MSCalendar.GetTimezone(c.user())
+	if err != nil {
+		return "Error: No timezone found", err
+	}
+
 	events, err := c.MSCalendar.ViewCalendar(c.user(), time.Now(), time.Now().Add(14*24*time.Hour))
 	if err != nil {
 		return "", err
 	}
 
-	return c.MSCalendar.RenderCalendarView(c.user(), events)
+	return views.RenderCalendarView(events, tz)
 }
