+++ server/mscalendar/calendar.go
@@ -4,8 +4,6 @@
 package mscalendar
 
 import (
-	"fmt"
-	"sort"
 	"time"
 
 	"github.com/mattermost/mattermost-plugin-mscalendar/server/remote"
@@ -18,7 +16,6 @@ type Calendar interface {
 	FindMeetingTimes(user *User, meetingParams *remote.FindMeetingTimesParameters) (*remote.MeetingTimeSuggestionResults, error)
 	GetCalendars(user *User) ([]*remote.Calendar, error)
 	ViewCalendar(user *User, from, to time.Time) ([]*remote.Event, error)
-	RenderCalendarView(user *User, events []*remote.Event) (string, error)
 }
 
 func (m *mscalendar) ViewCalendar(user *User, from, to time.Time) ([]*remote.Event, error) {
@@ -32,7 +29,7 @@ func (m *mscalendar) ViewCalendar(user *User, from, to time.Time) ([]*remote.Eve
 	return m.client.GetDefaultCalendarView(user.Remote.ID, from, to)
 }
 
-func (m *mscalendar) viewTodayCalendar(user *User) ([]*remote.Event, error) {
+func (m *mscalendar) viewTodayCalendar(user *User, timezone string) ([]*remote.Event, error) {
 	err := m.Filter(
 		withClient,
 	)
@@ -45,8 +42,11 @@ func (m *mscalendar) viewTodayCalendar(user *User) ([]*remote.Event, error) {
 		return nil, err
 	}
 
-	from := time.Now()
-	to := time.Now().Add(time.Hour * 24)
+	now := time.Now().UTC()
+	dt := remote.NewDateTime(now, "UTC").In(timezone)
+
+	from := dt.Time().Truncate(time.Hour * 24)
+	to := dt.Time().Add(time.Hour * 25)
 	return m.client.GetDefaultCalendarView(user.Remote.ID, from, to)
 }
 
@@ -118,74 +118,4 @@ func (m *mscalendar) GetCalendars(user *User) ([]*remote.Calendar, error) {
 	}
 
 	return m.client.GetCalendars(user.Remote.ID)
-}
-
-func (m *mscalendar) RenderCalendarView(user *User, events []*remote.Event) (string, error) {
-	if len(events) == 0 {
-		return "No events were found", nil
-	}
-
-	err := m.Filter(
-		withUserExpanded(user),
-	)
-	if err != nil {
-		return "", err
-	}
-
-	var timeZone string
-	tz, err := m.GetTimezone(user)
-	if err == nil {
-		timeZone = tz
-	}
-
-	if timeZone != "" {
-		for _, e := range events {
-			e.Start = e.Start.In(timeZone)
-			e.End = e.End.In(timeZone)
-		}
-	}
-
-	resp := "Times are shown in " + events[0].Start.TimeZone + "\n"
-	for _, group := range groupEventsByDate(events) {
-		resp += group[0].Start.Time().Format("Monday January 02") + "\n"
-		for _, e := range group {
-			resp += fmt.Sprintf("* %s\n", renderEvent(e))
-		}
-	}
-
-	return resp, nil
-}
-
-func renderEvent(event *remote.Event) string {
-	start := event.Start.Time().Format(time.Kitchen)
-	end := event.End.Time().Format(time.Kitchen)
-
-	return fmt.Sprintf("%s - %s `%s`", start, end, event.Subject)
-}
-
-func groupEventsByDate(events []*remote.Event) [][]*remote.Event {
-	groups := map[string][]*remote.Event{}
-
-	for _, event := range events {
-		date := event.Start.Time().Format("2006-01-02")
-		_, ok := groups[date]
-		if !ok {
-			groups[date] = []*remote.Event{}
-		}
-
-		groups[date] = append(groups[date], event)
-	}
-
-	days := []string{}
-	for k := range groups {
-		days = append(days, k)
-	}
-	sort.Strings(days)
-
-	result := [][]*remote.Event{}
-	for _, day := range days {
-		group := groups[day]
-		result = append(result, group)
-	}
-	return result
 }
